chore: log bridges move bencharks to bench_test.go files
